requirements.txt
fastmcp>=2.0.0
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.0
openpyxl>=3.1.0
rich>=13.7.0
